jeongyeop/crawler/requirements.txt
backports.zoneinfo==0.2.1;python_version<"3.9"
beautifulsoup4==4.12.2
Pillow==9.5.0
PyMySQL==1.0.2
scrapy==2.7.1
regex==2023.3.23
requests==2.29.0
selenium==4.9.0
webdriver-manager==3.8.6
tqdm==4.65.0
